[TC-10] 로그인, 회원가입 에러처리 수정
