+++ backend/README.md
@@ -1,0 +1,2 @@
+# Movie_test
+소공 영화추천 서비스 백엔드 test repo
