--- backend/gatchaPedia_BackEnd/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'demo'
