--- backend/gatchaPedia_BackEnd/src/main/java/com/gatchaPedia/demo/member/exception/DuplicateMemberByNameException.java
@@ -1,4 +1,0 @@
-package com.gatchaPedia.demo.member.exception;
-
-public class DuplicateMemberByNameException extends RuntimeException{
-}
