--- backend/gatchaPedia_BackEnd/src/main/resources/application.yml
@@ -1,20 +1,0 @@
-spring:
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/movie?useSSL=false&serverTimezone=Asia/Seoul&characterEncoding=UTF-8
-    username: root
-    password: worldcup7
-
-  jpa:
-    hibernate:
-      ddl-auto: update
-
-
-    properties:
-      hibernate:
-        format_sql: true
-
-
-logging.level:
-  org.hibernate.SQL: debug
-
