+++ backend/src/main/java/com/gatchaPedia/demo/member/exception/DuplicateMemberByUsernameException.java
