+++ backend/src/main/java/com/gatchaPedia/demo/member/response/SignUpInputFailResponse.java
@@ -6,13 +6,17 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.util.List;
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class LoginFailResponse {
+public class SignUpInputFailResponse {
 
     private Boolean success;
 
-    private String message;
+    private Long ERRCode;
+
+    private List<String> errorMessages;
 }
