--- backend/src/main/java/com/movie/demo/member/exception/PasswordMisMatchException.java
@@ -1,4 +1,0 @@
-package com.movie.demo.member.exception;
-
-public class PasswordMisMatchException extends RuntimeException{
-}
