--- backend/src/main/java/com/movie/demo/member/repository/MemberRepository.java
@@ -1,12 +1,0 @@
-package com.movie.demo.member.repository;
-
-import com.movie.demo.member.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-    boolean existsByUsername(String username);
-
-    Member findByUsername(String LoginId);
-
-}
