--- backend/src/main/java/com/movie/demo/member/request/LoginRequest.java
@@ -1,19 +1,0 @@
-package com.movie.demo.member.request;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class LoginRequest {
-
-    private String username;
-
-    private String password;
-
-}
