--- backend/src/main/java/com/movie/demo/member/request/SignUpRequest.java
@@ -1,20 +1,0 @@
-package com.movie.demo.member.request;
-
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class SignUpRequest {
-
-    private String name;
-
-    private String username;
-
-    private String password;
-
-    private String email;
-
-}
