--- backend/src/main/java/com/movie/demo/member/response/LoginResponse.java
@@ -1,20 +1,0 @@
-package com.movie.demo.member.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class LoginResponse {
-
-    private Boolean success;
-
-    private String message;
-
-    private Long memberId;
-
-}
