--- backend/src/main/java/com/movie/demo/member/response/SignUpResponse.java
@@ -1,19 +1,0 @@
-package com.movie.demo.member.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class SignUpResponse {
-
-    private Boolean success;
-
-    private String message;
-
-    private Long memberId;
-}
